separate functionality into authentication, shoppinglists and item using blue print
